fix: order building for Vaulting in AfterOrder (#192)
* fix: order building for Vaulting in AfterOrder

* rename test case
