+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 8.7.6
+- Fixes an issue, where vaulted payments could not be processed in the after order payment process in certain situations
+
 # 8.7.5
 - Fixes an issue with the pre-built storefront JavaScript (shopware/shopware#8483)
